// src/api/openapi/models/VehicleDocumentType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const VehicleDocumentType = {
    TECHNICAL_CERTIFICATE: 'TechnicalCertificate',
    INSURANCE: 'Insurance'
} as const;
export type VehicleDocumentType = typeof VehicleDocumentType[keyof typeof VehicleDocumentType];


export function instanceOfVehicleDocumentType(value: any): boolean {
    for (const key in VehicleDocumentType) {
        if (Object.prototype.hasOwnProperty.call(VehicleDocumentType, key)) {
            if (VehicleDocumentType[key as keyof typeof VehicleDocumentType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function VehicleDocumentTypeFromJSON(json: any): VehicleDocumentType {
    return VehicleDocumentTypeFromJSONTyped(json, false);
}

export function VehicleDocumentTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): VehicleDocumentType {
    return json as VehicleDocumentType;
}

export function VehicleDocumentTypeToJSON(value?: VehicleDocumentType | null): any {
    return value as any;
}

export function VehicleDocumentTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): VehicleDocumentType {
    return value as VehicleDocumentType;
}